Only clear root if it is non-empty.

+++ tests/megacmd_clear_root.sh
@@ -4,4 +4,9 @@ set -e
 
 mega-login $MEGA_EMAIL $MEGA_PWD
 
-mega-rm -rf "/*"
+LS_OUTPUT=`mega-ls "/" | xargs echo -n`
+if ["$LS_OUTPUT" = ""]; then
+    echo "Account cloud storage already empty, nothing to clear."
+else
+    mega-rm -rf "/*"
+fi
